+++ src/main/java/com/woorifisa/kboxwoori/domain/quiz/controller/QuizController.java
@@ -3,7 +3,6 @@ package com.woorifisa.kboxwoori.domain.quiz.controller;
 import com.woorifisa.kboxwoori.domain.quiz.dto.QuizRequestDto;
 import com.woorifisa.kboxwoori.domain.quiz.dto.QuizResponseDto;
 import com.woorifisa.kboxwoori.domain.quiz.dto.QuizResultResponseDto;
-import com.woorifisa.kboxwoori.domain.quiz.service.QuizParticipantService;
 import com.woorifisa.kboxwoori.domain.quiz.service.QuizService;
 import com.woorifisa.kboxwoori.domain.user.exception.NotAuthenticatedAccountException;
 import com.woorifisa.kboxwoori.global.config.security.PrincipalDetails;
@@ -20,13 +19,13 @@ import javax.validation.Valid;
 public class QuizController {
 
     private final QuizService quizService;
-    private final QuizParticipantService quizParticipantService;
 
     @GetMapping
     public ResponseDto<QuizResponseDto> getCurrentQuiz(@AuthenticationPrincipal PrincipalDetails principalDetails) {
         QuizResponseDto quiz = quizService.getCurrentQuiz();
+
         if (principalDetails != null) {
-            quiz.setHasParticipated(quizParticipantService.isUserParticipated(principalDetails.getUsername()));
+            quiz.setHasParticipated(quizService.isUserParticipated(principalDetails.getUsername()));
         }
         return ResponseDto.success(quiz);
     }
@@ -37,7 +36,6 @@ public class QuizController {
         if (principalDetails == null) {
             throw NotAuthenticatedAccountException.EXCEPTION;
         }
-
         return ResponseDto.success(quizService.submitAnswer(quizRequestDTO, principalDetails.getUsername()));
     }
 }
