+++ src/main/java/com/woorifisa/kboxwoori/domain/quiz/service/QuizService.java
@@ -1,13 +1,16 @@
 package com.woorifisa.kboxwoori.domain.quiz.service;
 
+import com.woorifisa.kboxwoori.domain.notification.service.NotificationService;
+import com.woorifisa.kboxwoori.domain.point.service.PointService;
 import com.woorifisa.kboxwoori.domain.quiz.dto.QuizRequestDto;
 import com.woorifisa.kboxwoori.domain.quiz.dto.QuizResponseDto;
 import com.woorifisa.kboxwoori.domain.quiz.dto.QuizResultResponseDto;
 import com.woorifisa.kboxwoori.domain.quiz.entity.Quiz;
-import com.woorifisa.kboxwoori.domain.quiz.exception.*;
+import com.woorifisa.kboxwoori.domain.quiz.exception.QuizDuplicatePartipation;
+import com.woorifisa.kboxwoori.domain.quiz.exception.QuizEndedException;
+import com.woorifisa.kboxwoori.domain.quiz.exception.QuizNotFoundException;
+import com.woorifisa.kboxwoori.domain.quiz.repository.QuizRedisRepository;
 import com.woorifisa.kboxwoori.domain.quiz.repository.QuizRepository;
-import com.woorifisa.kboxwoori.domain.user.entity.User;
-import com.woorifisa.kboxwoori.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,51 +22,39 @@ import java.util.List;
 public class QuizService {
 
     private final QuizRepository quizRepository;
-    private final UserRepository userRepository;
-    private final QuizParticipantService quizParticipantService;
+    private final QuizRedisRepository quizRedisRepository;
+    private final PointService pointService;
+    private final NotificationService notificationService;
 
     private static final int CORRECT_ANSWER_POINTS = 3;
 
     public QuizResponseDto getCurrentQuiz() {
-        //현재 날짜에 맞는 날짜
-        LocalDate currentDate = LocalDate.now();
-        //현재 날짜에 해당하는 퀴즈 가져옴
-        List <Quiz> quizzes = quizRepository.findByCreatedAt(currentDate);
+        List <Quiz> quizzes = quizRepository.findByCreatedAt(LocalDate.now());
 
         if (quizzes.isEmpty()) {
             throw QuizNotFoundException.EXCEPTION;
         }
 
-        QuizResponseDto quizResponseDTO = new QuizResponseDto(quizzes.get(0));
-        return quizResponseDTO;
+        return new QuizResponseDto(quizzes.get(0));
     }
 
     public QuizResultResponseDto submitAnswer(QuizRequestDto quizRequestDTO, String userId) {
-        LocalDate currentDate = LocalDate.now();
-
-        //주어진 quizId 퀴즈 가져옴
         Quiz quiz = quizRepository.findById(quizRequestDTO.getQuizId()).orElseThrow(() -> QuizNotFoundException.EXCEPTION);
 
-        if (!quiz.getCreatedAt().equals(currentDate)) {
+        if (!quiz.getCreatedAt().equals(LocalDate.now())) {
             throw QuizEndedException.EXCEPTION;
         }
 
-        User user = userRepository.findByUserId(userId).orElseThrow(() -> AccountNotFoundException.EXCEPTION);
-
         //사용자 참여 여부 redis 통해 확인
-        if (quizParticipantService.isUserParticipated(user.getUserId())) {
+        if (isUserParticipated(userId)) {
             throw QuizDuplicatePartipation.EXCEPTION;
         }
 
-        //정답 가져옴
-        char correctAnswer = quiz.getAnswer();
-
-        if (quizRequestDTO.getChoice() != correctAnswer) {
+        if (quizRequestDTO.getChoice() != quiz.getAnswer()) {
             return new QuizResultResponseDto(false);
         }
+
         //정답일 경우 처리
-        quizParticipantService.saveUserParticipation(user.getUserId());
-        //TODO: 포인트 적립
         quizRedisRepository.saveUserParticipation(userId);
         pointService.savePoint(userId, CORRECT_ANSWER_POINTS);
         notificationService.savePointNotification(userId, (long) CORRECT_ANSWER_POINTS);
@@ -71,5 +62,8 @@ public class QuizService {
         return new QuizResultResponseDto(true);
     }
 
-}
+    public Boolean isUserParticipated(String userId) {
+        return quizRedisRepository.isUserParticipated(userId);
+    }
 
+}
